backend/src/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
python-multipart
google-generativeai>=0.7.0
